runsc: correctly get the path for cgroup in container
The previous patch comparing `self` cgroup with `pid=1` cgroup, but
it's not semantically correct (they can be in different cgroup for
instance), and that's not the point as well, since we only need to
figure out the correct cgroup path for our pid.

Instead, we just need to compare the paths we have in `/proc/self/cgroup`
with `/proc/self/mountinfo` to get the cgroup path correctly.